requirements.txt
# Core Requirements
google-generativeai>=0.3.0
flask>=2.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
